==> 1. Yelp_API_Extractor.ipynb <==
import pandas as pd

Reading the Chicago food inspection csv file to dataframe

df=pd.read_csv("Food_Inspections.csv")

df

Extracting only the restaurants name and the addresses

df[["DBA Name","Address"]]

Dropping the duplicate restaurants and addresses pairs later to be used for hitting the yelp API

df_dup_removed=df[["DBA Name","Address"]].drop_duplicates()

df_dup_removed

Exporting the unique restaurants names and their addresses dataframe to a csv file

df_dup_removed.to_csv('ChicagoRestaurantsAndAddress.csv', index = False)

import json
import requests

Making a url string that will store the yelp api and a list of API keys to increase the total number of hits that can be done per day to the yelp API as there is a certain limit of hits per API keys.

url = 'https://api.yelp.com/v3/businesses/matches'

# List of API keys commented for security reasons

#api_keys = ["####"]

Reading the unique restaurants names and addresses pairs csv to a dataframe and hitting the API for each pairs. For each unique restaurant and addresses pairs 2 API hits are done. First one to extract the business IDs from the yelp api which is stored in the yelp.json file. Second hit is done by using this business ID to get the business details which is stored in yelp.json file.

rest_df = pd.read_csv('data/ChicagoRestaurantsAndAddress.csv')
rest = rest_df.shape[0]

offset = 0
for i, api_key in enumerate(api_keys):
    headers = {'Authorization': 'Bearer %s' % api_key}
        
    j = 0
    while j in range(2450) and offset <= rest - 1:
        params = {'name': rest_df.iloc[offset]['DBA Name'], 'location': 'Chicago', 'address1': rest_df.iloc[offset]['Address'], 'city': 'Chicago', 'state': 'IL', 'country': 'US', 'match_threshold': 'default'}
        
        req = requests.get(url, params = params, headers = headers)
        data = json.loads(req.text)
        offset += 1
        j += 1

        if(len(data['businesses']) > 0):
            with open("data/yelp.json", "r+") as file:
                data_list = json.load(file)
                data_list.append(data)
                file.seek(0)
                json.dump(data_list, file)

            id_ = data['businesses'][0]['id']
            url_2 = 'https://api.yelp.com/v3/businesses/{}'.format(id_)
            req_2 = requests.get(url_2, headers = headers)
            data_2 = json.loads(req_2.text)

            with open("data/restaurant.json", "r+") as file:
                data_list = json.load(file)
                data_list.append(data_2)
                file.seek(0)
                json.dump(data_list, file)

print(offset)
