==> 2. Similarity_Matching.ipynb <==
import pandas as pd

Reading the restaurant.json file to dfapi dataframe

dfapi=pd.read_json('restaurant.json')

dfapi

## Creating a list of names and addresses of the restaurants that were read from the yelp.json file

# creating a list which will store address 1 
add1_list=[]
for i in range(len(dfapi)):
    add1_list.append(dfapi['location'][i]['display_address'][0])
    
#creating a list which will store only the names of the restaurants
name_list=list(dfapi['name'])

### Merging the restaurants name list with their address list created above

#creating a list which will store both names and address 1
name_add1_api=[]
for i in range(len(add1_list)):
    name_add1_api.append([name_list[i],add1_list[i]])
name_add1_api

### Creating a function similar() using SequenceMatcher method of difflib library for finding the similarity between the two strings, the function similar() will return 1 if both the strings completely match and 0 when both there is no similarit between both the strings

from difflib import SequenceMatcher

def similar(a, b):
    return SequenceMatcher(None, a, b).ratio()

### Fetching the unique names and addresses data of restaurants from the unique restaurants names and addresses csv that was created in the first notebook

dfinsp=pd.read_csv('ChicagoRestaurantsAndAddress.csv')

### Finding the matching indexes for both the data by similarity

similarity_name=0.50
similarity_address=0.65
similar_index=[]
start_search=0
for i in range(len(name_add1_api)):     
    print(i)
    similar_index.append(-1)  #keeping the default value as -1 which will give us the indication for when similar index not found
   
    for j in (list(range(start_search,len(dfinsp)))+ list(range(start_search))): #starting the search from the last found index
        if pd.isna(dfinsp['DBA Name'][j]):
            continue
        if similar(name_add1_api[i][0].lower(),dfinsp['DBA Name'][j].lower())>similarity_name and similar(name_add1_api[i][1].lower(),dfinsp['Address'][j].lower())>similarity_address:
            similar_index[i]=j
            start_search=j+1  # next search to begin after jth row
            break


### Finding the matching index for both the data by using fuzzywuzzy library

from fuzzywuzzy import fuzz
from fuzzywuzzy import process

similarity_name=80
similarity_address=98
similar_index=[]
start_search=0

for i in range(len(name_add1_api)):  #inititalizing all the similar_index to -1 
  similar_index.append(-1)      #keeping the default value as -1 which will give us the indication if similar index not found

while similarity_address>=70:
  start_search=0                          #to start the search from 0 for every iteration of similarity_index
  for i in range(len(name_add1_api)):     #starting the search from the last found index                 
      print(i,similarity_address)         #printing this data for keeping track of the progress of the loop
      if(similar_index[i]==-1):  #the below search should only run if similar index has not be found
        for j in (list(range(start_search,len(dfinsp)))+ list(range(start_search))):
            if pd.isna(dfinsp['DBA Name'][j]):
                continue
            if fuzz.partial_ratio(name_add1_api[i][0].lower(),dfinsp['DBA Name'][j].lower())>similarity_name and fuzz.partial_ratio(name_add1_api[i][1].lower(),dfinsp['Address'][j].lower())>similarity_address:
                similar_index[i]=j
                start_search=j+1  # next search to begin after jth row
                break
  similarity_address=similarity_address-5
  if(similarity_name>=45):          #name similarity should not go less than 40 percent
    similarity_name=similarity_name-5

### Better matching was done by the fuzzywuzzy library so the matching done by the difflib library was discarded as it was less accurate. All the matched pairs of indexes from both the dataset were stored in a csv file

pd.DataFrame(similar_index).to_csv("Similar_index_fuzzy_all_iteration.csv")

### Above code was run in google collab to save high processing time and the output was exported to a csv (Similar_index_fuzzy_all_iteration.csv) file which is imported below

Similar_index=pd.read_csv('Similar_index_fuzzy_all_iteration.csv')

Similar_index.columns


Renaming the columns after importing the csv file. YELP_API column containing the index of the data that was fetched from the YELP API and the INSPECTION column which stores the index of the Chicago Inspection dataset that was matched to the corresponding YELP_API index.

Similar_index=Similar_index.rename(columns={'Unnamed: 0': "YELP_API", '0': "INSPECTION"})

### Extracting the restaurant names and addresses from the index data

match_name=[]
match_address=[]
for i in range(len(Similar_index)):
    if Similar_index['INSPECTION'][i]==-1:
        match_name.append('Not Found')
        match_address.append('Not Found')
    else:
        match_name.append(dfinsp["DBA Name"][Similar_index['INSPECTION'][i]])
        match_address.append(dfinsp["Address"][Similar_index['INSPECTION'][i]])
    

finalpd=pd.DataFrame(name_add1_api)
finalpd=finalpd.rename(columns={'0': "YELP_Name", '1': "Yelp_Address"})
finalpd['InspName']=match_name
finalpd['InspAddress']=match_address

'finalpd' dataframe contains the names and addresses of the restaurants that we were fetched from the YELP API in the first two columns and the next two columns contains the names and addresses of the restaurants from the CHICAGO FOOD INSPECTION Dataset that were matched to the corresponding YELP API data by using fuzzywuzzy library

finalpd

### Exporting this dataframe to a csv file

pd.DataFrame(finalpd).to_csv("finaldataframe_fuzzy_alliterations.csv")
