==> 3. Dataset_Merging.ipynb <==
import pandas as pd

Reading the Chicago Food Inspection dataset to a dataframe

inspections = pd.read_csv('Food_Inspections.csv')

inspections.head(10)

Extracting the year of inspectiong from the Inspection Date column

inspections["Year"] = inspections["Inspection Date"].str.strip().str.split('/').str[-1]
inspections["Year"] = pd.to_numeric(inspections["Year"])
inspections.head()

years = list(inspections["Year"])
type(years[0])

Sorting the data by inspection year and dropping all the previous years records

inspections.sort_values(by = "Year", ascending = False, inplace = True)

inspections.drop_duplicates(subset = ["DBA Name", "Address"], inplace = True)

inspections.shape[0]

inspections["Year"].value_counts()

Storing this data to a csv so that we won't have to do the above steps repeatedly 

inspections.to_csv('Inspection with latest records and duplicates removed.csv')

Reading the above created csv file in a new dataframe

dfinsp=pd.read_csv("Inspection with latest records and duplicates removed.csv")

dfinsp

Reading the matched YELP API and Chicago Food Inspection data from a csv file created in the "2.DS matching the two data sets.ipynb" notebook 

dfmatch=pd.read_csv("finaldataframe_fuzzy_alliterations.csv")

dfmatch

Reading the Chicago Food Inspection data in a dataframe

dfyelp=pd.read_json("restaurant.json")

Merging all the columns fetched by YELP API with the dfmatch index and storing it into df_merge_YELP dataframe

df_merge_YELP=pd.concat([dfmatch, dfyelp], axis=1)

df_merge_YELP

Creating a new column which will store the concatenation of the restaurants names and their addresses which will be used for merging in dfinsp dataframe

name_add_string_list=[]
for i in range(len(dfinsp)):
    name_str=str(dfinsp['DBA Name'][i])
    add_str=str(dfinsp['Address'][i])
    name_add_string_list.append( name_str+ add_str)
dfinsp['NameAddString']=name_add_string_list
    

dfinsp

Doing similar thing for df_merge_YELP dataframe


name_add_string_list=[]
for i in range(len(df_merge_YELP)):
    name_str=str(df_merge_YELP['InspName'][i])
    add_str=str(df_merge_YELP['InspAddress'][i])
    name_add_string_list.append( name_str+ add_str)
df_merge_YELP['NameAddString']=name_add_string_list
    

df_merge_YELP

Using this concatenated restaurants names and address to merge both the dataset and store this data in df_final_merge dataframe

df_final_merge=pd.merge(df_merge_YELP,dfinsp,on='NameAddString')

df_final_merge

Dropping duplicates from the df_final_merge dataframe

df_final_merge=df_final_merge.drop_duplicates(['0','1'])

Renaming the columns

df_final_merge.rename(columns={'0':'Yelp_Name','1':'Yelp_Address'})

Both of our datasets are completely merged and we export this final dataset to a csv file on which all the analysis will be done in the "4.Analysis_Clustering_Visualization.ipynb" notebook

df_final_merge.to_csv("Both dataset merged.csv")
